x264enc: fix unsigned comparison warning

+++ ext/x264/gstx264enc.c
@@ -1964,7 +1964,7 @@ gst_x264_enc_encode_frame (GstX264Enc * encoder, x264_picture_t * pic_in,
     }
   }
 
-  if (pic_out.i_dts + encoder->dts_offset < 0) {
+  if (pic_out.i_dts + (gint64) encoder->dts_offset < 0) {
     /* should be ok now, surprise if not */
     GST_WARNING_OBJECT (encoder, "negative dts after offset compensation");
     frame->dts = GST_CLOCK_TIME_NONE;
